Split up core into small modules
