+++ src/test/lib/unsafe-html_test.ts
@@ -15,7 +15,8 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
-import {html, render} from '../../core.js';
+import {html} from '../../index.js';
+import {render} from '../../lib/render.js';
 import {unsafeHTML} from '../../lib/unsafe-html.js';
 import {stripExpressionDelimeters} from '../test-helpers.js';
